backend/requirements.txt
amqp==5.3.1
asgiref==3.7.2
async-timeout==5.0.1
attrs==23.2.0
billiard==4.2.1
celery==5.4.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.5.1
ddsketch==2.0.4
ddtrace==1.1.3
Django==4.2.10
django-cors-headers==4.3.1
django-countries==4.6.1
django-ipware==2.1.0
django-user-agents==0.3.2
djangorestframework==3.14.0
geoip2==2.9.0
idna==3.3
ipaddress==1.0.22
kombu==5.4.2
maxminddb==2.2.0
mssql-django==1.0rc1
packaging==24.0
prompt_toolkit==3.0.50
protobuf==5.26.0
psycopg2-binary==2.9.10
pygame==2.6.0
pyodbc==5.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
qualname==0.1.0
redis==5.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
typing_extensions==4.12.2
tzdata==2025.1
ua-parser==0.16.1
urllib3==1.22
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.5
